requirements.txt
reportlab
python-barcode